moved new data to dummyData
