+++ server/data/dummyData.ts
@@ -1,4 +1,5 @@
-import { GithubStatusList } from "../types/shared"
+import { CommitSummary } from "../types/CommitSummary";
+import { GithubStatusList } from "../types/shared";
 
 export const githubStatus: GithubStatusList = [
   {
@@ -28,4 +29,38 @@ export const githubStatus: GithubStatusList = [
     branch: "main",
     pass: true,
   },
-]
+];
+
+const now = new Date();
+
+export const sampleCommitSummary: CommitSummary = {
+  user: "dxren",
+  time: now,
+  message: "integrated slack bot wooooh",
+  filesChanged: 1,
+  linesAdded: 55,
+  linesRemoved: 66,
+};
+
+const sampleCommitSummary2: CommitSummary = {
+  user: "yabochk8",
+  time: now,
+  message: "broke all my tests and had to start again",
+  filesChanged: 2,
+  linesAdded: 77,
+  linesRemoved: 166,
+};
+
+const sampleCommitSummary3: CommitSummary = {
+  user: "dxren",
+  time: now,
+  message: "turns out the slack bot",
+  filesChanged: 4,
+  linesAdded: 55,
+  linesRemoved: 66,
+};
+export const sampleCommitSummaries = [
+  sampleCommitSummary,
+  sampleCommitSummary2,
+  sampleCommitSummary3,
+];
