+++ server/types/CommitSummary.ts
@@ -1,0 +1,9 @@
+export type CommitSummary = {
+  user: string;
+  time: Date;
+  message: string;
+  filesChanged: number;
+  linesAdded: number;
+  linesRemoved: number;
+  actualChanges?: string;
+};
